requirements.txt
streamlit
langchain
langchain-huggingface
transformers
torch
accelerate
python-dotenv
